requirements-local.txt
requests
flask
flask-rest-api
mongoengine
pandas
pymongo
tensorflow
transformers